fix export_to_phy template_ind
If compute_pc_features was False templates_ind was wrong. 
Because it used pc_feature_ind and the new method for calculate template ind without pca_ind was wrong (work with template without all channels). 
So I use channel_list return when waveform are exctracted that appear to be the same as pc_feature_ind. 

I don't remove the old method l 1462 but maybe we can?