export_to_phy add bad_channel_ids argument
The function don't work if you provide recording without bad channels, so I add it as argument.
The issue with bad channels is that it could be take as best_channel when it's not, only because there is aberrant high value. 
So I template of bad channels to 0 when best channel is calculated.

Did the way I do it is ok for you?