Create artifact suppression preprocessor
This is a simple preprocessor that I've been using on my data. It looks for unreasonably high amplitude events, and removes them, filling with either zeros, or noise matched to the data. Sometimes, with very long chronic recordings, the tether can get a bit loose, and there will be noise artifacts or motion artifacts in the data that mess with different sorting algorithms.  I've been using it in my chronic recordings and it helps in salvaging recordings that have small sections of corrupted data.